requirements.txt
python-dotenv
pandas
numpy
scikit-learn
pymongo
certifi
dill
pyaml

# -e .
